Refactoring / relocating a few files that should not be in the root of /src folder
