+++ src/classes/animation_presets.py
