+++ src/tests/test_dialog_preview_resize.py
@@ -16,7 +16,7 @@ if PATH not in sys.path:
 
 TEST_MEDIA_ROOT = os.path.join(os.sep, "mock-media")
 
-from qt_test_app import ensure_app_state, get_or_create_app
+from tests.qt_test_app import ensure_app_state, get_or_create_app
 
 
 class DummySettings:
