+++ src/tests/test_main_window.py
@@ -50,7 +50,7 @@ from qt_api import QApplication, QDockWidget, QMainWindow, QMenu, QStandardItem,
 from classes.project_data import ProjectDataStore
 from classes.settings import SettingStore
 from classes.updates import UpdateManager
-from qt_test_app import ensure_app_state as ensure_qt_app_state, get_or_create_app
+from tests.qt_test_app import ensure_app_state as ensure_qt_app_state, get_or_create_app
 
 QCoreApplication.setAttribute(Qt.AA_ShareOpenGLContexts, True)
