+++ src/tests/test_timeline_helpers.py
@@ -47,7 +47,7 @@ from qt_api import QColor, QCursor, QImage, QPainter
 from qt_api import QApplication
 from classes import info
 from classes.updates import UpdateAction
-from qt_test_app import ensure_app_state as ensure_qt_app_state, get_or_create_app
+from tests.qt_test_app import ensure_app_state as ensure_qt_app_state, get_or_create_app
 
 QCoreApplication.setAttribute(Qt.AA_ShareOpenGLContexts, True)
