+++ src/tests/test_video_widget_transform.py
@@ -40,7 +40,7 @@ PATH = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 if PATH not in sys.path:
     sys.path.append(PATH)
 
-from qt_test_app import ensure_app_state, get_or_create_app
+from tests.qt_test_app import ensure_app_state, get_or_create_app
 
 
 class DummySettings:
